+++ src/pkg/common/getWorkspacePath.js
@@ -1,5 +1,7 @@
 const vscode = require("vscode");
-
+/**
+ * Returns the current local root folder's address
+ */
 async function getWorkspacePath() {
   const workspaceFolders = vscode.workspace.workspaceFolders;
   if (workspaceFolders && workspaceFolders.length > 0) {