+++ src/pkg/formats/formatReadme.js
@@ -1,10 +1,17 @@
+/**
+ * Formats the log contents for teh auto-generated readme file
+ * @param {string[]} - Changes from the `git diff --stat` cmd
+ * @param {string} - date
+ * @param {string} - startTime
+ * @param {string} - endTime
+ */
 function formatReadme(changes, date, startTime, endTime) {
   const changesArray = changes.trim().split("\n");
 
   const totalSummary = changesArray.pop();
 
   const logContent = `
-**⚠️ Warning:** Avoid interacting with the \`project-history/*\` folder and \`cache.txt\` file directly.
+**⚠️ Warning:** Avoid interacting with the \`work-logs/*\` folder and \`cache.txt\` file directly.
 
 # Project Logs