+++ src/pkg/formats/formatTime.js
@@ -1,3 +1,8 @@
+/**
+ * Formats the Time to this format `11:00 PM/AM`
+ * @param {Date} - date
+ * @param {string} returns - formattedTime
+ */
 function formatTime(date) {
   const formattedTime = date.toLocaleString("en-US", {
     hour: "numeric",